refactor: change package
add product domain, global package
