+++ app/src/main/java/com/codesoom/assignment/product/controllers/ProductController.java
@@ -1,8 +1,8 @@
-package com.codesoom.assignment.controllers;
+package com.codesoom.assignment.product.controllers;
 
-import com.codesoom.assignment.application.ProductService;
-import com.codesoom.assignment.domain.Product;
-import com.codesoom.assignment.dto.ProductData;
+import com.codesoom.assignment.product.application.ProductService;
+import com.codesoom.assignment.product.domain.Product;
+import com.codesoom.assignment.product.dto.ProductData;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
