--- app/src/test/java/com/codesoom/assignment/controllers/UserControllerTest.java
@@ -1,4 +1,0 @@
-package com.codesoom.assignment.controllers;
-
-public class UserControllerTest {
-}
