+++ app/src/test/java/com/codesoom/assignment/health/controllers/HelloControllerTest.java
@@ -1,5 +1,6 @@
-package com.codesoom.assignment.controllers;
+package com.codesoom.assignment.health.controllers;
 
+import com.codesoom.assignment.health.controllers.HelloController;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
